refactor(BetMoneyTest): 베팅금액을 계산한 결과를 정수로 가져오도록 변형

+++ src/main/java/participant/player/BetMoney.java
@@ -18,7 +18,7 @@ public class BetMoney {
         }
     }
 
-    public double betMoneyResult(double percent) {
-        return betAmount * percent;
+    public int betMoneyResult(double percent) {
+        return (int) (betAmount * percent);
     }
 }
